+++ examples/delegated-routing/src/index.js
@@ -1,8 +1,0 @@
-// eslint-disable-next-line
-'use strict'
-
-import React from 'react' // eslint-disable-line no-unused-vars
-import ReactDOM from 'react-dom'
-import App from './App.js' // eslint-disable-line no-unused-vars
-
-ReactDOM.render(<App />, document.getElementById('root'))