+++ examples/libp2p-in-the-browser/webrtc/browser-to-server/vite.config.js
@@ -1,8 +1,0 @@
-export default {
-  build: {
-    target: 'es2022'
-  },
-  optimizeDeps: {
-    esbuildOptions: { target: 'es2022', supported: { bigint: true } }
-  },
-}