+++ packages/kad-dht/test/routing-table.spec.ts
@@ -1,6 +1,6 @@
 /* eslint-env mocha */
 
-import { EventEmitter, CustomEvent } from '@libp2p/interface/events'
+import { TypedEventEmitter, CustomEvent } from '@libp2p/interface/events'
 import { mockConnectionManager } from '@libp2p/interface-compliance-tests/mocks'
 import { PeerSet } from '@libp2p/peer-collections'
 import { peerIdFromString } from '@libp2p/peer-id'
@@ -33,7 +33,7 @@ describe('Routing Table', () => {
   beforeEach(async function () {
     this.timeout(20 * 1000)
 
-    const events = new EventEmitter<Libp2pEvents>()
+    const events = new TypedEventEmitter<Libp2pEvents>()
 
     components = {
       peerId: await createPeerId(),