+++ packages/libp2p/test/transports/transport-manager.spec.ts
@@ -1,6 +1,6 @@
 /* eslint-env mocha */
 
-import { EventEmitter } from '@libp2p/interface/events'
+import { TypedEventEmitter } from '@libp2p/interface/events'
 import { start, stop } from '@libp2p/interface/startable'
 import { FaultTolerance } from '@libp2p/interface/transport'
 import { mockUpgrader } from '@libp2p/interface-compliance-tests/mocks'
@@ -26,7 +26,7 @@ describe('Transport Manager (WebSockets)', () => {
   let components: Components
 
   beforeEach(async () => {
-    const events = new EventEmitter()
+    const events = new TypedEventEmitter()
     components = {
       peerId: await createEd25519PeerId(),
       events,